Overhaul formatting of Graphite check failures.
This gives better detail on which specific series failed the checks
which can help wildcard checks be more informative.

To save space, we strip the nodes that all series have in common from
the front and back of the series targets, so "prefix.a.suffix" and
"prefix.b.suffix" turn into just "a" and "b".